+++ tbsim/__init__.py
@@ -1,7 +1,7 @@
 import os
 DATADIR = os.path.join(os.path.dirname(__file__), 'data')
 
-from .tb_lshtm import *
+from .tb import *
 from .comorbidities import *
 from .interventions import *
 from .networks import *